# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product to the cart",
                "description": "This endpoint adds a product to the user's cart. If the product is already in the cart, it increases the quantity. If the stock is insufficient, it returns an error.",
                "operationId": "af5c3a2c065e9b55aaa49e4cf9b0f9ab",
                "requestBody": {
                    "description": "Data for adding a product to the cart",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity",
                                    "customer_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the product to add to the cart",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Quantity of the product to add",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "customer_id": {
                                        "description": "ID of the customer",
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product added to cart successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product added to cart successfully"
                                        },
                                        "cart": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "customer_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-11-13T10:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-11-13T10:05:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Insufficient Stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No Stock"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Out of Stock."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "messages": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected product_id is invalid."
                                                    }
                                                },
                                                "quantity": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The quantity field must be an integer."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/view": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "View all items in the cart",
                "description": "This endpoint returns all the products added to the customer's cart. If the cart is empty, a 404 response is returned.",
                "operationId": "062ced901100f2e3380e2c6d30208775",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved cart items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cart_items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "productId": {
                                                        "description": "ID of the product in the cart",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "customer_id": {
                                                        "description": "ID of the customer",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Product name",
                                                        "type": "string",
                                                        "example": "Sample Product"
                                                    },
                                                    "id": {
                                                        "description": "Product ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "cartId": {
                                                        "description": "Cart item ID",
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "quantity": {
                                                        "description": "Quantity of the product in the cart",
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cart is empty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your cart is empty."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}